Fixed 500 error when POST new Student with same first and last name
Previously a default Security User was being created every time a new Student is created. If the new Student has the same First and Last name as a previously created Student then attempting to build the Security User would throw Constraint Violation Exception since the default username for Security User uses the first and last name to create a default username. This is a Unique column in Security User.

The change was made in buildSecurityUserForNewStudent method. Now we verify that the proposed default username is not already present in the database. If it isn't then we go ahead and make the SecurityUser for the new Student. If it is then the new Security User is not generated for the new Student.